requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
Django==4.1.6
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
idna==3.4
Pillow==9.4.0
psycopg2==2.9.5
pytz==2022.7.1
requests==2.28.2
sqlparse==0.4.3
urllib3==1.26.14
uvicorn==0.20.0
